# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "WebNewsAPI",
        "contact": {}
    },
    "paths": {
        "/news": {
            "post": {
                "description": "Create a news for the app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "Create News",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newsHandler.CreateNewsObject"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.CreateNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "Delete news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.DeleteNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/newsHandler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "newsHandler.CreateNewsObject": {
            "type": "object",
            "properties": {
                "firstContent": {
                    "type": "string"
                },
                "secondContent": {
                    "type": "string"
                },
                "thirdContent": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "newsHandler.CreateNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.DeleteNewsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/schema.NewsResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "newsHandler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "schema.NewsResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "firstContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "secondContent": {
                    "type": "string"
                },
                "thirdContent": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}